This is vAmiga v2.1b4

+++ Emulator/Agnus/Sequencer/Sequencer.h
@@ -318,7 +318,7 @@ private:
     void updateBplJumpTable(i16 end = HPOS_MAX);
 
     // Computes the layout of a single fetch unit
-    void computeFetchUnit(u8 dmacon);
+    void computeFetchUnit(u16 dmacon);
     template <u8 channels> void computeLoresFetchUnit();
     template <u8 channels> void computeHiresFetchUnit();
     template <u8 channels> void computeShresFetchUnit();
